refactor: move interface to card component
