--- react-components/src/pages/Home/components/card/interfaces.tsx
@@ -1,8 +1,0 @@
-export interface ICard {
-  img: string;
-  name: string;
-  description: string;
-  text: string;
-  likes: number;
-  date: string;
-}
